requirements.txt
PyQt5
QtPy
h5py
umap-learn
seaborn
matplotlib
numpy
scikit-learn
scipy
similaritymeasures

